🚚 Rename find_workspaces -> switch_workspace
