+++ lua/telescope/_extensions/neorg.lua
@@ -8,6 +8,6 @@ return require("telescope").register_extension({
         find_project_tasks = require("neorg.modules.core.integrations.telescope.module").public.find_project_tasks,
         find_aof_tasks = require("neorg.modules.core.integrations.telescope.module").public.find_aof_tasks,
         find_aof_project_tasks = require("neorg.modules.core.integrations.telescope.module").public.find_aof_project_tasks,
-        find_workspaces = require("neorg.modules.core.integrations.telescope.module").public.find_workspaces
+        switch_workspace = require("neorg.modules.core.integrations.telescope.module").public.switch_workspace
     },
 })
